Refactor namespaces to organize sender-related classes under Ratatosk.Senders
