+++ src/Ratatosk.Senders/Senders/ISenderCache.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for details.
 //
 
-namespace Ratatosk
+namespace Ratatosk.Senders
 {
     /// <summary>
     /// Caches sender identities by name to avoid repeated lookups in the registry.
