+++ src/Ratatosk.Senders/Senders/SenderBuilder.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for details.
 //
 
-namespace Ratatosk
+namespace Ratatosk.Senders
 {
     /// <summary>
     /// Builds a <see cref="Sender"/> instance using a fluent API.
