+++ src/Ratatosk.Senders/Senders/SenderConnectorBuilder.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for details.
 //
 
-namespace Ratatosk
+namespace Ratatosk.Senders
 {
     /// <summary>
     /// Configures sender resolution for a specific connector type.
