+++ src/Ratatosk.Senders/Senders/SenderConnectorOptions.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for details.
 //
 
-namespace Ratatosk
+namespace Ratatosk.Senders
 {
     /// <summary>
     /// Options for per-connector sender resolution configuration.
