+++ src/Ratatosk/SenderRegistrationBuilder.cs
@@ -3,6 +3,8 @@
 // Licensed under the MIT license. See LICENSE file in the project root for details.
 //
 
+using Ratatosk.Senders;
+
 namespace Ratatosk
 {
     /// <summary>
