+++ test/Ratatosk.Senders.XUnit/Unit/InMemorySenderCacheTests.cs
@@ -1,4 +1,4 @@
-namespace Ratatosk;
+namespace Ratatosk.Senders;
 
 [Trait("Category", "Unit")]
 [Trait("Layer", "Domain")]
