+++ test/Ratatosk.Senders.XUnit/Unit/SenderValidatorTests.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace Ratatosk;
+namespace Ratatosk.Senders;
 
 [Trait("Category", "Unit")]
 [Trait("Layer", "Domain")]
